// src/core/enums.ts
export enum Activity {
  Absent = 0,
  Minimal = 1,
  Few = 2,
  Some = 3,
  Moderate = 4,
  Many = 5,
  Abundant = 6,
  Swarms = 7,
  NA = 100,
}
// ...
export enum EquipmentType {
  Weapon = 0,
  Shield = 1,
  Gadget = 2,
}
// ...
export enum ShipType {
  Flea = 0,
  Gnat = 1,
  Firefly = 2,
  Mosquito = 3,
  Bumblebee = 4,
  Beetle = 5,
  Hornet = 6,
  Grasshopper = 7,
  Termite = 8,
  Wasp = 9,
  SpaceMonster = 10,
  Dragonfly = 11,
  Mantis = 12,
  Scarab = 13,
  Bottle = 14,
  Custom = 15,
  Scorpion = 16,
}

export enum ShipyardId {
  NA = -1,
  Corellian = 0,
  Incom = 1,
  Kuat = 2,
  Sienar = 3,
  Sorosuub = 4,
}

export enum ShipyardSkill {
  CrewQuarters = 0,
  FuelBase = 1,
  HullPerUnit = 2,
  ShieldSlotUnits = 3,
  WeaponSlotUnits = 4,
}

export enum Size {
  Tiny = 0,
  Small = 1,
  Medium = 2,
  Large = 3,
  Huge = 4,
  Gargantuan = 5,
}
// ...
export enum TechLevel {
  PreAgricultural = 0,
  Agricultural = 1,
  Medieval = 2,
  Renaissance = 3,
  EarlyIndustrial = 4,
  Industrial = 5,
  PostIndustrial = 6,
  HiTech = 7,
  Unavailable = 8,
}